fix default watch directories
Morbo doesn't watch the current directory. It watches the 'lib' and
'templates' directories and the Mojolicious app script. We guess at the
app script if we can, and pick the 'lib' and 'templates' directories.

Watching the current directory by default also ensures that anything
written to the `MOJO_HOME` will cause a restart. Since it's common for
things to be written here, this is a bad default.